add script
