+++ public/index.html
@@ -16,6 +16,6 @@
     <div id="observe"></div>
   </div>
 </body>
-<script type="module" src="src/index.js"></script>
+<script type="script src="../src/index.js"></script>
 
 </html>
