+++ src/index.js
@@ -1,4 +1,18 @@
-import Data from "src/Data.js";
+class Data {
+  static saveData(data = {}, name = "") {
+    localStorage.setItem(name, JSON.stringify(data));
+    //console.log(localStorage);
+  }
+  static loadData(name = "") {
+    return JSON.parse(localStorage.getItem(name));
+  }
+  static removeData(name = "") {
+    localStorage.removeItem(name);
+  }
+  static clear() {
+    localStorage.clear();
+  }
+}
 
 const $app = document.getElementById("app");
 const $observe = document.getElementById("observe");
